refactor(stats): drop dead IPC handlers, unify stats types over HTTP con
- Remove unused stats data channels/handlers from ipc.ts; stats now flows only over HTTP (preload stays limited to native window behavior)
- Add src/types/stats-wire.ts + stats-http-contract.ts as the single source of truth for stats payloads; stats-server.ts wraps responses via statsJson()
- Delete stats/src/lib/ipc-client.ts; update api-client.ts/mining.ts/stats.ts to consume the unified contract types
- Add stats-transport-architecture.test.ts to guard the HTTP-only boundary; update docs/architecture notes accordingly
