+++ src/core/services/ipc.test.ts
@@ -181,35 +181,6 @@ function createFakeImmersionTracker(
 ): NonNullable<IpcServiceDeps['immersionTracker']> {
   return {
     recordYomitanLookup: () => {},
-    getSessionSummaries: async () => [],
-    getDailyRollups: async () => [],
-    getMonthlyRollups: async () => [],
-    getQueryHints: async () => ({
-      totalSessions: 0,
-      activeSessions: 0,
-      episodesToday: 0,
-      activeAnimeCount: 0,
-      totalActiveMin: 0,
-      totalCards: 0,
-      activeDays: 0,
-      totalEpisodesWatched: 0,
-      totalAnimeCompleted: 0,
-      totalTokensSeen: 0,
-      totalLookupCount: 0,
-      totalLookupHits: 0,
-      totalYomitanLookupCount: 0,
-      newWordsToday: 0,
-      newWordsThisWeek: 0,
-    }),
-    getSessionTimeline: async () => [],
-    getSessionEvents: async () => [],
-    getVocabularyStats: async () => [],
-    getKanjiStats: async () => [],
-    getMediaLibrary: async () => [],
-    getMediaDetail: async () => null,
-    getMediaSessions: async () => [],
-    getMediaDailyRollups: async () => [],
-    getCoverArt: async () => null,
     markActiveVideoWatched: async () => false,
     ...overrides,
   };
@@ -774,172 +745,6 @@ test('registerIpcHandlers records yomitan lookup when subtitle context recording
   } finally {
     console.warn = originalWarn;
   }
-});
-
-test('registerIpcHandlers returns empty stats overview shape without a tracker', async () => {
-  const { registrar, handlers } = createFakeIpcRegistrar();
-  registerIpcHandlers(createRegisterIpcDeps(), registrar);
-
-  const overviewHandler = handlers.handle.get(IPC_CHANNELS.request.statsGetOverview);
-  assert.ok(overviewHandler);
-  assert.deepEqual(await overviewHandler!({}), {
-    sessions: [],
-    rollups: [],
-    hints: {
-      totalSessions: 0,
-      activeSessions: 0,
-      episodesToday: 0,
-      activeAnimeCount: 0,
-      totalCards: 0,
-      totalActiveMin: 0,
-      activeDays: 0,
-      totalEpisodesWatched: 0,
-      totalAnimeCompleted: 0,
-      totalTokensSeen: 0,
-      totalLookupCount: 0,
-      totalLookupHits: 0,
-      totalYomitanLookupCount: 0,
-      newWordsToday: 0,
-      newWordsThisWeek: 0,
-    },
-  });
-});
-
-test('registerIpcHandlers validates and clamps stats request limits', async () => {
-  const { registrar, handlers } = createFakeIpcRegistrar();
-  const calls: Array<[string, number, number?]> = [];
-
-  registerIpcHandlers(
-    createRegisterIpcDeps({
-      immersionTracker: {
-        recordYomitanLookup: () => {},
-        getSessionSummaries: async (limit = 0) => {
-          calls.push(['sessions', limit]);
-          return [];
-        },
-        getDailyRollups: async (limit = 0) => {
-          calls.push(['daily', limit]);
-          return [];
-        },
-        getMonthlyRollups: async (limit = 0) => {
-          calls.push(['monthly', limit]);
-          return [];
-        },
-        getQueryHints: async () => ({
-          totalSessions: 0,
-          activeSessions: 0,
-          episodesToday: 0,
-          activeAnimeCount: 0,
-          totalCards: 0,
-          totalActiveMin: 0,
-          activeDays: 0,
-          totalEpisodesWatched: 0,
-          totalAnimeCompleted: 0,
-          totalTokensSeen: 0,
-          totalLookupCount: 0,
-          totalLookupHits: 0,
-          totalYomitanLookupCount: 0,
-          newWordsToday: 0,
-          newWordsThisWeek: 0,
-        }),
-        getSessionTimeline: async (sessionId: number, limit = 0) => {
-          calls.push(['timeline', limit, sessionId]);
-          return [];
-        },
-        getSessionEvents: async (sessionId: number, limit = 0) => {
-          calls.push(['events', limit, sessionId]);
-          return [];
-        },
-        getVocabularyStats: async (limit = 0) => {
-          calls.push(['vocabulary', limit]);
-          return [];
-        },
-        getKanjiStats: async (limit = 0) => {
-          calls.push(['kanji', limit]);
-          return [];
-        },
-        getMediaLibrary: async () => [],
-        getMediaDetail: async () => null,
-        getMediaSessions: async () => [],
-        getMediaDailyRollups: async () => [],
-        getCoverArt: async () => null,
-        markActiveVideoWatched: async () => false,
-      },
-    }),
-    registrar,
-  );
-
-  await handlers.handle.get(IPC_CHANNELS.request.statsGetDailyRollups)!({}, -1);
-  await handlers.handle.get(IPC_CHANNELS.request.statsGetMonthlyRollups)!(
-    {},
-    Number.POSITIVE_INFINITY,
-  );
-  await handlers.handle.get(IPC_CHANNELS.request.statsGetSessions)!({}, 9999);
-  await handlers.handle.get(IPC_CHANNELS.request.statsGetSessionTimeline)!({}, 7, 12.5);
-  await handlers.handle.get(IPC_CHANNELS.request.statsGetSessionEvents)!({}, 7, 0);
-  await handlers.handle.get(IPC_CHANNELS.request.statsGetVocabulary)!({}, 1000);
-  await handlers.handle.get(IPC_CHANNELS.request.statsGetKanji)!({}, NaN);
-
-  assert.deepEqual(calls, [
-    ['daily', 60],
-    ['monthly', 24],
-    ['sessions', 500],
-    ['timeline', 200, 7],
-    ['events', 500, 7],
-    ['vocabulary', 500],
-    ['kanji', 100],
-  ]);
-});
-
-test('registerIpcHandlers requests the full timeline when no limit is provided', async () => {
-  const { registrar, handlers } = createFakeIpcRegistrar();
-  const calls: Array<[string, number | undefined, number]> = [];
-
-  registerIpcHandlers(
-    createRegisterIpcDeps({
-      immersionTracker: {
-        recordYomitanLookup: () => {},
-        getSessionSummaries: async () => [],
-        getDailyRollups: async () => [],
-        getMonthlyRollups: async () => [],
-        getQueryHints: async () => ({
-          totalSessions: 0,
-          activeSessions: 0,
-          episodesToday: 0,
-          activeAnimeCount: 0,
-          totalCards: 0,
-          totalActiveMin: 0,
-          activeDays: 0,
-          totalEpisodesWatched: 0,
-          totalAnimeCompleted: 0,
-          totalTokensSeen: 0,
-          totalLookupCount: 0,
-          totalLookupHits: 0,
-          totalYomitanLookupCount: 0,
-          newWordsToday: 0,
-          newWordsThisWeek: 0,
-        }),
-        getSessionTimeline: async (sessionId: number, limit?: number) => {
-          calls.push(['timeline', limit, sessionId]);
-          return [];
-        },
-        getSessionEvents: async () => [],
-        getVocabularyStats: async () => [],
-        getKanjiStats: async () => [],
-        getMediaLibrary: async () => [],
-        getMediaDetail: async () => null,
-        getMediaSessions: async () => [],
-        getMediaDailyRollups: async () => [],
-        getCoverArt: async () => null,
-        markActiveVideoWatched: async () => false,
-      },
-    }),
-    registrar,
-  );
-
-  await handlers.handle.get(IPC_CHANNELS.request.statsGetSessionTimeline)!({}, 7, undefined);
-
-  assert.deepEqual(calls, [['timeline', undefined, 7]]);
 });
 
 test('registerIpcHandlers ignores malformed fire-and-forget payloads', () => {
