+++ src/core/services/ipc.ts
@@ -132,35 +132,6 @@ export interface IpcServiceDeps {
   ) => Promise<PlaylistBrowserMutationResult>;
   immersionTracker?: {
     recordYomitanLookup: () => void;
-    getSessionSummaries: (limit?: number) => Promise<unknown>;
-    getDailyRollups: (limit?: number) => Promise<unknown>;
-    getMonthlyRollups: (limit?: number) => Promise<unknown>;
-    getQueryHints: () => Promise<{
-      totalSessions: number;
-      activeSessions: number;
-      episodesToday: number;
-      activeAnimeCount: number;
-      totalActiveMin: number;
-      totalCards: number;
-      activeDays: number;
-      totalEpisodesWatched: number;
-      totalAnimeCompleted: number;
-      totalTokensSeen: number;
-      totalLookupCount: number;
-      totalLookupHits: number;
-      totalYomitanLookupCount: number;
-      newWordsToday: number;
-      newWordsThisWeek: number;
-    }>;
-    getSessionTimeline: (sessionId: number, limit?: number) => Promise<unknown>;
-    getSessionEvents: (sessionId: number, limit?: number) => Promise<unknown>;
-    getVocabularyStats: (limit?: number) => Promise<unknown>;
-    getKanjiStats: (limit?: number) => Promise<unknown>;
-    getMediaLibrary: () => Promise<unknown>;
-    getMediaDetail: (videoId: number) => Promise<unknown>;
-    getMediaSessions: (videoId: number, limit?: number) => Promise<unknown>;
-    getMediaDailyRollups: (videoId: number, limit?: number) => Promise<unknown>;
-    getCoverArt: (videoId: number) => Promise<unknown>;
     markActiveVideoWatched: () => Promise<boolean>;
   } | null;
 }
@@ -459,24 +430,6 @@ export function createIpcDepsRuntime(options: IpcDepsRuntimeOptions): IpcService
 }
 
 export function registerIpcHandlers(deps: IpcServiceDeps, ipc: IpcMainRegistrar = ipcMain): void {
-  const parsePositiveIntLimit = (
-    value: unknown,
-    defaultValue: number,
-    maxValue: number,
-  ): number => {
-    if (!Number.isInteger(value) || (value as number) < 1) {
-      return defaultValue;
-    }
-    return Math.min(value as number, maxValue);
-  };
-
-  const parsePositiveInteger = (value: unknown): number | null => {
-    if (typeof value !== 'number' || !Number.isInteger(value) || value <= 0) {
-      return null;
-    }
-    return value;
-  };
-
   ipc.on(
     IPC_CHANNELS.command.setIgnoreMouseEvents,
     (event: unknown, ignore: unknown, options: unknown = {}) => {
@@ -904,115 +857,4 @@ export function registerIpcHandlers(deps: IpcServiceDeps, ipc: IpcMainRegistrar 
       return await deps.movePlaylistBrowserIndex(index as number, direction as 1 | -1);
     },
   );
-
-  // Stats request handlers
-  ipc.handle(IPC_CHANNELS.request.statsGetOverview, async () => {
-    const tracker = deps.immersionTracker;
-    if (!tracker) {
-      return {
-        sessions: [],
-        rollups: [],
-        hints: {
-          totalSessions: 0,
-          activeSessions: 0,
-          episodesToday: 0,
-          activeAnimeCount: 0,
-          totalActiveMin: 0,
-          totalCards: 0,
-          activeDays: 0,
-          totalEpisodesWatched: 0,
-          totalAnimeCompleted: 0,
-          totalTokensSeen: 0,
-          totalLookupCount: 0,
-          totalLookupHits: 0,
-          totalYomitanLookupCount: 0,
-          newWordsToday: 0,
-          newWordsThisWeek: 0,
-        },
-      };
-    }
-    const [sessions, rollups, hints] = await Promise.all([
-      tracker.getSessionSummaries(5),
-      tracker.getDailyRollups(14),
-      tracker.getQueryHints(),
-    ]);
-    return { sessions, rollups, hints };
-  });
-
-  ipc.handle(IPC_CHANNELS.request.statsGetDailyRollups, async (_event, limit: unknown) => {
-    const parsedLimit = parsePositiveIntLimit(limit, 60, 500);
-    return deps.immersionTracker?.getDailyRollups(parsedLimit) ?? [];
-  });
-
-  ipc.handle(IPC_CHANNELS.request.statsGetMonthlyRollups, async (_event, limit: unknown) => {
-    const parsedLimit = parsePositiveIntLimit(limit, 24, 120);
-    return deps.immersionTracker?.getMonthlyRollups(parsedLimit) ?? [];
-  });
-
-  ipc.handle(IPC_CHANNELS.request.statsGetSessions, async (_event, limit: unknown) => {
-    const parsedLimit = parsePositiveIntLimit(limit, 50, 500);
-    return deps.immersionTracker?.getSessionSummaries(parsedLimit) ?? [];
-  });
-
-  ipc.handle(
-    IPC_CHANNELS.request.statsGetSessionTimeline,
-    async (_event, sessionId: unknown, limit: unknown) => {
-      const parsedSessionId = parsePositiveInteger(sessionId);
-      if (parsedSessionId === null) return [];
-      const parsedLimit = limit === undefined ? undefined : parsePositiveIntLimit(limit, 200, 1000);
-      return deps.immersionTracker?.getSessionTimeline(parsedSessionId, parsedLimit) ?? [];
-    },
-  );
-
-  ipc.handle(
-    IPC_CHANNELS.request.statsGetSessionEvents,
-    async (_event, sessionId: unknown, limit: unknown) => {
-      const parsedSessionId = parsePositiveInteger(sessionId);
-      if (parsedSessionId === null) return [];
-      const parsedLimit = parsePositiveIntLimit(limit, 500, 1000);
-      return deps.immersionTracker?.getSessionEvents(parsedSessionId, parsedLimit) ?? [];
-    },
-  );
-
-  ipc.handle(IPC_CHANNELS.request.statsGetVocabulary, async (_event, limit: unknown) => {
-    const parsedLimit = parsePositiveIntLimit(limit, 100, 500);
-    return deps.immersionTracker?.getVocabularyStats(parsedLimit) ?? [];
-  });
-
-  ipc.handle(IPC_CHANNELS.request.statsGetKanji, async (_event, limit: unknown) => {
-    const parsedLimit = parsePositiveIntLimit(limit, 100, 500);
-    return deps.immersionTracker?.getKanjiStats(parsedLimit) ?? [];
-  });
-
-  ipc.handle(IPC_CHANNELS.request.statsGetMediaLibrary, async () => {
-    return deps.immersionTracker?.getMediaLibrary() ?? [];
-  });
-
-  ipc.handle(IPC_CHANNELS.request.statsGetMediaDetail, async (_event, videoId: unknown) => {
-    if (typeof videoId !== 'number') return null;
-    return deps.immersionTracker?.getMediaDetail(videoId) ?? null;
-  });
-
-  ipc.handle(
-    IPC_CHANNELS.request.statsGetMediaSessions,
-    async (_event, videoId: unknown, limit: unknown) => {
-      if (typeof videoId !== 'number') return [];
-      const parsedLimit = parsePositiveIntLimit(limit, 100, 500);
-      return deps.immersionTracker?.getMediaSessions(videoId, parsedLimit) ?? [];
-    },
-  );
-
-  ipc.handle(
-    IPC_CHANNELS.request.statsGetMediaDailyRollups,
-    async (_event, videoId: unknown, limit: unknown) => {
-      if (typeof videoId !== 'number') return [];
-      const parsedLimit = parsePositiveIntLimit(limit, 90, 500);
-      return deps.immersionTracker?.getMediaDailyRollups(videoId, parsedLimit) ?? [];
-    },
-  );
-
-  ipc.handle(IPC_CHANNELS.request.statsGetMediaCover, async (_event, videoId: unknown) => {
-    if (typeof videoId !== 'number') return null;
-    return deps.immersionTracker?.getCoverArt(videoId) ?? null;
-  });
 }
