+++ src/shared/ipc/contracts.ts
@@ -101,19 +101,6 @@ export const IPC_CHANNELS = {
     animetoshoDownloadFile: 'animetosho:download-file',
     animetoshoGetSecondaryLanguages: 'animetosho:get-secondary-languages',
     kikuBuildMergePreview: 'kiku:build-merge-preview',
-    statsGetOverview: 'stats:get-overview',
-    statsGetDailyRollups: 'stats:get-daily-rollups',
-    statsGetMonthlyRollups: 'stats:get-monthly-rollups',
-    statsGetSessions: 'stats:get-sessions',
-    statsGetSessionTimeline: 'stats:get-session-timeline',
-    statsGetSessionEvents: 'stats:get-session-events',
-    statsGetVocabulary: 'stats:get-vocabulary',
-    statsGetKanji: 'stats:get-kanji',
-    statsGetMediaLibrary: 'stats:get-media-library',
-    statsGetMediaDetail: 'stats:get-media-detail',
-    statsGetMediaSessions: 'stats:get-media-sessions',
-    statsGetMediaDailyRollups: 'stats:get-media-daily-rollups',
-    statsGetMediaCover: 'stats:get-media-cover',
     getConfigSettingsSnapshot: 'config:get-settings-snapshot',
     saveConfigSettingsPatch: 'config:save-settings-patch',
     openConfigSettingsFile: 'config:open-settings-file',
